+++ packages/cadl-ranch-coverage-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @azure-tools/cadl-ranch-api
 
+## 0.5.0
+
+### Minor Changes
+
+- d0f5be0: bump tsp deps to 0.50.0
+
+## 0.4.0
+
+### Minor Changes
+
+- 2896d45: Update to consume tsp 49
+
+## 0.3.1
+
+### Patch Changes
+
+- ddb62ac: General dependency update
+
 ## 0.3.0
 
 ### Minor Changes